+++ packages/mint-components/src/components/sqm-dropdown-field/sqm-dropdown-field.tsx
@@ -101,7 +101,7 @@ function useDropdownFieldDemo(
   return deepmerge(
     {
       states: {
-        validationErrors: [],
+        registrationFormState: {},
       },
     },
     props.demoData || {},